Box office values should be floats.

+++ sources/processors/wikipedia/rank.py
@@ -45,7 +45,7 @@ class WikipediaRankProcessor(RankProcessor):
     def box_office(page, wikidata):
         box_office_property = "P2142"
         return next(
-            (int(claim["value"]["amount"]) for claim in wikidata.get("claims", [])
+            (float(claim["value"]["amount"]) for claim in wikidata.get("claims", [])
              if claim["property"] == box_office_property)
             , 0)
